+++ NEWS.md
@@ -4,6 +4,12 @@ This page describes the most important changes in `TypeUtils`. The format is bas
 [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to
 [Semantic Versioning](https://semver.org/spec).
 
+## Unreleased
+
+### Added
+
+- Extend `similar` for the type of arrays returned by `as_eltype`.
+
 ## Version 1.6.0 (2025-04-15)
 
 ### Added